결과 출력에 대해 클래스 통합
