+++ src/main/java/racingcar/view/PrintOutput.java
@@ -1,6 +1,7 @@
 package racingcar.view;
 
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 public class PrintOutput {
@@ -22,5 +23,14 @@ public class PrintOutput {
             printMove(item.getValue());
         }
         System.out.println();
+    }
+
+    public static void printWinner(List<String> winner_list) {
+        System.out.print("최종 우승자 : ");
+
+        System.out.print(winner_list.get(0));
+        for (int i = 1; i < winner_list.size(); i++) {
+            System.out.print(", " + winner_list.get(i));
+        }
     }
 }
